src/libtest.py
import pandas as pd
import numpy as np
import os
import yaml
import joblib
import mlflow
from sklearn.model_selection import train_test_split
from sksurv.ensemble import RandomSurvivalForest
from sksurv.metrics import concordance_index_censored
print("Все библиотеки загружены")