requirements.txt
joblib==1.5.2
mlflow==3.6.0
numpy==1.24.2
pandas==2.3.3
PyYAML==6.0.3
scikit_learn==1.7.2
scikit_survival==0.25.0
dvc>=3.0.0